pyproject.toml
[project]
name = "llm-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "litellm>=1.80.9",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.1",
]
